refactor: combined rest lambdas into one binary
